requirements.txt
opencv-python
mysql-connector-python
ultralytics
paddleocr
imutils
python-dotenv